add filter_other_outcomes(...)
Filter away instances with other outcomes. If an instances has both an admissible outcome and a non-admissible one, it will still be kept. Maybe we should remove the non-admissible tag from the document? It is not a problem for the graph representation algorithm (we simply discard them if needed) so for the moment let's keep it as it is.